# docs/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Everruns API",
    "description": "API for managing AI agents, sessions, messages, and events",
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    },
    "version": "0.2.0"
  },
  "servers": [
    {
      "url": "https://app.everruns.com/api",
      "description": "Production API"
    }
  ],
  "paths": {
...
    "/v1/images": {
      "get": {
        "tags": [
          "images"
        ],
        "summary": "GET /v1/images - List images",
        "operationId": "list_images",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "description": "Max items to return (default 50)",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "description": "Items to skip",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of images",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ImageInfo"
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal server error"
          }
        }
      },
      "post": {
        "tags": [
          "images"
        ],
        "summary": "POST /v1/images - Upload an image",
        "operationId": "upload_image",
        "parameters": [
          {
            "name": "session_id",
            "in": "query",
            "description": "Optional: session ID stored as metadata for tracking (not required for upload)",
            "required": false,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "description": "Multipart form data with 'file' field",
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "string"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Image uploaded successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ImageUploadResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request (bad format, size exceeded, etc.)"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/v1/images/{image_id}": {
      "get": {
        "tags": [
          "images"
        ],
        "summary": "GET /v1/images/{image_id} - Get image (returns binary data)",
        "operationId": "get_image",
        "parameters": [
          {
            "name": "image_id",
            "in": "path",
            "description": "Image ID (prefixed, e.g., img_...)",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Image binary data",
            "content": {
              "image/*": {}
            }
          },
          "400": {
            "description": "Invalid image ID"
          },
          "404": {
            "description": "Image not found"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      },
      "delete": {
        "tags": [
          "images"
        ],
        "summary": "DELETE /v1/images/{image_id} - Delete an image",
        "operationId": "delete_image",
        "parameters": [
          {
            "name": "image_id",
            "in": "path",
            "description": "Image ID (prefixed, e.g., img_...)",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Image deleted"
          },
          "400": {
            "description": "Invalid image ID"
          },
          "404": {
            "description": "Image not found"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/v1/images/{image_id}/thumbnail": {
      "get": {
        "tags": [
          "images"
        ],
        "summary": "GET /v1/images/{image_id}/thumbnail - Get image thumbnail",
        "operationId": "get_thumbnail",
        "parameters": [
          {
            "name": "image_id",
            "in": "path",
            "description": "Image ID (prefixed, e.g., img_...)",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Thumbnail binary data",
            "content": {
              "image/jpeg": {}
            }
          },
          "400": {
            "description": "Invalid image ID"
          },
          "404": {
            "description": "Image not found or no thumbnail"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ImageInfo": {
        "type": "object",
        "description": "Image metadata (without binary data)",
        "required": [
          "id",
          "filename",
          "content_type",
          "size_bytes",
          "metadata",
          "created_at"
        ],
        "properties": {
          "content_type": {
            "type": "string"
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          },
          "filename": {
            "type": "string"
          },
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "metadata": {},
          "size_bytes": {
            "type": "integer",
            "format": "int64"
          }
        }
      },
      "ImageUploadResponse": {
        "type": "object",
        "description": "Image upload response",
        "required": [
          "id",
          "filename",
          "content_type",
          "size_bytes",
          "created_at"
        ],
        "properties": {
          "content_type": {
            "type": "string"
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          },
          "filename": {
            "type": "string"
          },
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "size_bytes": {
            "type": "integer",
            "format": "int64"
          }
        }
      },
...
    }
  },
...
}